Simplifying Horizon Image
Removed the need for a opennms-data directory which adds unnecessary complexity. Just mount the data directories for
rrd, reports and mibs if you need them.

Refactored the combination of update/init and run functions. The entrypoint allowed -s and -n to have a combination of
doing an upgrade / or init and start Horizon w/o Cassandra. The functions are removed and we just provide distinct
operational functions which just do one thing, initialize the configs or do an upgrade and start Horizon.

We provided multiple overlay directories for
  * opennms/etc -> /opt/opennms-etc-overlay
  * opennms/jetty-webapp/opennms/WEB-INF -> /opt/opennms-jetty-webinf-overlay
  * opennms -> /opt/opennms-overlay

The latest introduced opennms-overlay superseeds the existing ones, to remove complexity we remove the
opennms-overlay directory and remove the other ones.